pyproject.toml
[project]
name = "sia-lightning-langgraph"
version = "0.1.0"
description = "Self-Improving Agentic System Using Langgraph"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain",
    "langgraph",
    "langchain_openai",
    "tavily-python",
    "datasets", 
    "wandb",
    "pandas", 
    "python-dotenv",
    "rich"
]
